LPS-199467 move methods from MediaWikiImporter to the only usage on WikiNodeLocalServiceImpl, and delete the other class
